[MISC] Change test from int to long. Otherwise it gives incorrect answer when test is for example as big as 1e12.

+++ GEAR/src/gear/subcommands/metawatchdog/encrypt/EnigmaCommandImpl.java
@@ -79,7 +79,7 @@ public class EnigmaCommandImpl extends CommandImpl
 		double K = file.readDouble();
 		seed = file.readLong();
 		double alpha = file.readDouble();
-		int tests = file.readInt();
+		long tests = file.readInt();
 
 		double beta = file.readDouble();
 		double b = file.readDouble();
